creates class UnifiedMultiAgentTransitions
